+++ tests/commands/test_organizations.py
@@ -17,6 +17,7 @@
 # with Crate these terms will supersede the license and you may use the
 # software solely pursuant to the terms of the relevant commercial agreement.
 
+import uuid
 from unittest import mock
 
 import pytest
@@ -132,6 +133,13 @@ def test_organizations_edit_no_arguments(mock_request, capsys):
 def test_organizations_list(mock_request):
     call_command("croud", "organizations", "list")
     assert_rest(mock_request, RequestMethod.GET, "/api/v2/organizations/")
+
+
+@mock.patch.object(Client, "request", return_value=({}, None))
+def test_organizations_get(mock_request):
+    id = str(uuid.uuid4())
+    call_command("croud", "organizations", "get", id)
+    assert_rest(mock_request, RequestMethod.GET, f"/api/v2/organizations/{id}/")
 
 
 @mock.patch.object(Client, "request", return_value=(None, {}))