+++ tests/commands/test_subscriptions.py
@@ -17,6 +17,7 @@
 # with Crate these terms will supersede the license and you may use the
 # software solely pursuant to the terms of the relevant commercial agreement.
 
+import uuid
 from unittest import mock
 
 from croud.api import Client, RequestMethod
@@ -27,3 +28,10 @@ from tests.util import assert_rest, call_command
 def test_subscriptions_list(mock_request):
     call_command("croud", "subscriptions", "list")
     assert_rest(mock_request, RequestMethod.GET, "/api/v2/subscriptions/")
+
+
+@mock.patch.object(Client, "request", return_value=({}, None))
+def test_subscriptions_get(mock_request):
+    id = str(uuid.uuid4())
+    call_command("croud", "subscriptions", "get", id)
+    assert_rest(mock_request, RequestMethod.GET, f"/api/v2/subscriptions/{id}/")